% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzUninfinite.R
\name{adjust_padding}
\alias{adjust_padding}
\title{Adjust padding for Infinite coordinates}
\usage{
adjust_padding(coord_values, padding)
}
\arguments{
\item{coord_values}{numeric. The coordinate x and y values, potentially
containing Inf or -Inf}

\item{padding}{Result of \code{\link{get_padding}}}
}
\value{
A character vector of adjusted padding lengths.
}
\description{
Add appropriate sign to the padding values, depending on the sign of the
infinity.
}
\details{
\itemize{
\item -Inf has to be shifted in the positive direction
\item +Inf has to be shifted in the negative direction
}
}
