% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzUninfinite.R
\name{discretize}
\alias{discretize}
\title{Replace Infinites by discrete values}
\usage{
discretize(coord_values, xrange, yrange)
}
\arguments{
\item{coord_values}{numeric. The coordinate x and y values, potentially
containing Inf or -Inf}

\item{xrange}{Numeric vector of length 2, minimum and maximum values in the x
direction}

\item{yrange}{Numeric vector of length 2, minimum and maximum values in the y
direction}
}
\description{
The replacement values correspond to the edges of the available coordinate
space
}
