% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzUninfinite.R
\name{get_padding}
\alias{get_padding}
\title{Calculate length of padding from plot borders}
\usage{
get_padding(gg_plot)
}
\arguments{
\item{gg_plot}{A ggplot2 object.}
}
\value{
A character vector of paddings for \code{t}, \code{r}, \code{l}, \code{b}.
}
\description{
To prevent overlap with panel borders or axis lines, Infinites in TikZ
coordinates are padded with the thickness of these lines. They depend on the
current plot theme.
}
\seealso{
\code{\link{uninfinite_coord}} for construction of the complete
replaced coordinate.
}
