% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzCanvas.R
\name{gg_to_npc.ggtikzCanvas}
\alias{gg_to_npc.ggtikzCanvas}
\title{Convert data coordinates to npc coordinates.}
\usage{
\method{gg_to_npc}{ggtikzCanvas}(self, coord, panelx, panely)
}
\arguments{
\item{self}{a \code{\link{ggtikzCanvas}} object}

\item{coord}{A numeric vector of length 2, with the x coordinate to convert
at \code{coord[1]} and the y coordinate to convert at \code{coord[2]}}

\item{panelx}{X position (column) of the panel holding the data}

\item{panely}{X position (row) of the panel holding the data}
}
\value{
The input coordinates from \code{coord} converted to npc coordinates in
the form of a numeric vector of length 2. (0,0) corresponds to the lower
left corner of the viewport containing the \code{ggplot} panel specified by
\code{panelx} and \code{panely}, and (1,1) corresponds to the upper right corner.
}
\description{
Convert data coordinates to npc coordinates.
}
