% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzCanvas.R
\name{ggtikzCanvas}
\alias{ggtikzCanvas}
\title{Create a canvas to store TikZ annotations to a ggplot.}
\usage{
ggtikzCanvas(gg_plot)
}
\arguments{
\item{gg_plot}{A ggplot object on which annotations should be made.}
}
\value{
A ggtikz object, to which annotations can be added.
}
\description{
Annotations can be made relative to the whole plot, to a panel, or to data
coordinates (of individual panels).
}
\details{
This function provides a canvas for TikZ annotations, and does not draw
anything by itself. Its purpose is to provide information about the underlying
ggplot object for coordinate calculations.
}
\seealso{
\code{\link[tikzDevice]{grid.tikzAnnotate}} for annotation of base graphics.

\code{\link{ggtikz}} for a helper function for quick one-step annotations.

\code{\link{ggtikzAnnotation}} for more information about creating and adding ggtikz annotations.
}
