% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzUninfinite.R
\name{ggtikzUninfinite}
\alias{ggtikzUninfinite}
\title{Replace Inf in TikZ coordinates}
\usage{
ggtikzUninfinite(ggtikzCanvas, ggtikzAnnotation)
}
\arguments{
\item{ggtikzCanvas}{A \code{link{ggtikzCanvas}} object.}

\item{ggtikzAnnotation}{A \code{link{ggtikzAnnotaton}} object.}
}
\value{
A \code{link{ggtikzAnnotation}} object, with Infinites in coordinates
replaced by finite values.
}
\description{
Infinite values in TiKZ coordinate specifications are replaced by values
corresponding to the edge of the available coordinate space. This allows
placement of annotations at the very edge of a panel without knowing its
precise coordinates. This is useful for annotations which extend
to the panel boundaries, but also make use of specific coordinates.
}
