% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclip.R
\name{set_ggtikz_unclip_hook}
\alias{set_ggtikz_unclip_hook}
\alias{unset_ggtikz_unclip_hook}
\title{Unclip plots produced by the \code{tikzDevice}.}
\usage{
set_ggtikz_unclip_hook()

unset_ggtikz_unclip_hook()
}
\value{
Called for side effects - the \code{unclip} knitr hook is set or unset,
respectively.
}
\description{
By default, plots produced with the tikzDevice are clipped to the plot area,
which also clips ggtikzAnnotations extending beyond the plot boundaries. This
function removes the 'clip' and 'use as bounding box' options in a tikz file.
}
\seealso{
\code{\link{unclip}}, the hook that is being set.
}
