% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_utils.R
\name{split_coord}
\alias{split_coord}
\title{Split a TikZ coordinate.}
\usage{
split_coord(coord)
}
\arguments{
\item{coord}{Coordinate string of the form "(x,y)"}
}
\value{
A character vector of length 2: The x and y components of the
coordinate. These may contain spaces.
}
\description{
Split a TikZ coordinate.
}
