% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclip.R
\name{tikz_exts_pattern}
\alias{tikz_exts_pattern}
\title{Construct a regex pattern for possible tikzDevice extensions.}
\usage{
tikz_exts_pattern(options)
}
\description{
Construct a regex pattern for possible tikzDevice extensions.
}
