% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclip.R
\name{unclip}
\alias{unclip}
\title{knitr hook to remove clipping from plots produced with the tikzDevice.}
\usage{
unclip(before, options)
}
\arguments{
\item{before}{see \code{\link[knitr]{knit_hooks}}}

\item{options}{see \code{\link[knitr]{knit_hooks}}}
}
\value{
Called for side effect.
The files containing tikz plots are edited and overwritten.
}
\description{
Note that the chunk options \code{unclip = TRUE} and \code{external = FALSE}must be set
for the hook to come into effect!
}
\seealso{
\code{\link{set_ggtikz_unclip_hook}} to set the knitr hook.

\code{\link{unclip_tikz}}, the workhorse function for this hook.
}
