% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzUninfinite.R
\name{uninfinite_coord}
\alias{uninfinite_coord}
\alias{uninfinite_tikz}
\title{Replace infinite values in TikZ coordinates}
\usage{
uninfinite_coord(coord, xrange, yrange)

uninfinite_tikz(tikz_code, xrange, yrange)
}
\arguments{
\item{coord}{TikZ coordinate}

\item{xrange}{Numeric vector of length 2, minimum and maximum values in the x
direction}

\item{yrange}{Numeric vector of length 2, minimum and maximum values in the y
direction}

\item{tikz_code}{The TikZ code to replace Infinite values in.}
}
\value{
The adjusted TikZ coordinate with padding, as a string.
}
\description{
Infinite values are replaced with the minimum or maximum value of the padding
in the x or y direction, respectively. Additionally, the adjusted coordinate
is padded so that it lies just next to the panel borders and axis lines
without overlap.
}
