% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtikzCanvas.R
\name{gg_to_npc.ggtikzCanvas}
\alias{gg_to_npc.ggtikzCanvas}
\title{Convert data coordinates to npc coordinates.}
\usage{
\method{gg_to_npc}{ggtikzCanvas}(self, coord, panelx, panely)
}
\arguments{
\item{self}{a \code{\link{ggtikzCanvas}} object}

\item{coord}{A numeric vector of length two, with the x coordinate to convert
at \code{coord[1]} and the y coordinate to convert at \code{coord[2]}}

\item{panelx}{X positition (column) of the panel holding the data}

\item{panely}{X positition (row) of the panel holding the data}
}
\description{
Convert data coordinates to npc coordinates.
}
