% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclip.R
\name{tikz_exts_pattern}
\alias{tikz_exts_pattern}
\title{Construct a regex pattern for possible tikzDevice extensions.}
\usage{
tikz_exts_pattern(options)
}
\arguments{
\item{options}{A list of \code{knitr} chunk options}
}
\value{
A regex pattern to match file extensions of tikz figures
}
\description{
Construct a regex pattern for possible tikzDevice extensions.
}
